feature: move screen files with composable functions to package screen
Put all the files containing composable functions inside a package called screens.
Put also in the LazyColumn of the entries screen the rememberLazyListState function in order to remember the lazy list state of the list.

Integrates #11
