+++ app/src/main/java/com/example/mycity/ui/screens/Details.kt
@@ -1,4 +1,4 @@
-package com.example.mycity.ui
+package com.example.mycity.ui.screens
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Arrangement
